+++ lib/predicate/nodes/subset.rb
@@ -1,10 +1,10 @@
 class Predicate
-  module Intersect
+  module Subset
     include SetOp
 
     def evaluate(tuple)
       x, y = left.evaluate(tuple), right.evaluate(tuple)
-      x && y && !(x & y).empty?
+      x && y && (x & y == x)
     end
 
   end